// data/github/current/ui/packages/sub-issues/components/__generated__/SubIssuesListStoryQuery.graphql.ts
export type SubIssuesListStoryQuery$variables = Record<PropertyKey, never>;
export type SubIssuesListStoryQuery$data = {
  readonly node: {
    readonly " $fragmentSpreads": FragmentRefs<"SubIssuesList">;
  } | null | undefined;
  readonly viewer: {
    readonly " $fragmentSpreads": FragmentRefs<"SubIssuesListViewViewer">;
  };
};
export type SubIssuesListStoryQuery = {
  response: SubIssuesListStoryQuery$data;
  variables: SubIssuesListStoryQuery$variables;
};